Create Cabal-described
Move Described functionality there, out of Cabal the library.
Similarly cabal-install doesn't contain Described instances,
they are only in tests.
