+++ Cabal/Cabal-described/Cabal-described.cabal
@@ -1,0 +1,26 @@
+cabal-version: 2.2
+name:          Cabal-described
+version:       3.3.0.0
+synopsis:      Described functionality for types in Cabal
+category:      Testing, Parsec
+description:   Provides rere bindings
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  ghc-options:      -Wall
+  build-depends:
+    , base
+    , Cabal             ^>=3.3.0.0
+    , containers
+    , pretty
+    , QuickCheck
+    , rere              ^>=0.1
+    , tasty
+    , tasty-quickcheck
+    , transformers
+
+  exposed-modules:
+    Distribution.Described
+    Distribution.Utils.CharSet
+    Distribution.Utils.GrammarRegex
