+++ Cabal/Distribution/FieldGrammar/Newtypes.hs
@@ -38,14 +38,13 @@ import Distribution.Compat.Prelude
 import Prelude ()
 
 import Distribution.CabalSpecVersion
-import Distribution.Compiler               (CompilerFlavor)
-import Distribution.FieldGrammar.Described
-import Distribution.License                (License)
+import Distribution.Compiler         (CompilerFlavor)
+import Distribution.License          (License)
 import Distribution.Parsec
 import Distribution.Pretty
 import Distribution.Version
-       (LowerBound (..), Version, VersionRange, VersionRangeF (..), cataVersionRange, anyVersion, asVersionIntervals, mkVersion, version0, versionNumbers)
-import Text.PrettyPrint                    (Doc, comma, fsep, punctuate, vcat, (<+>), text)
+       (LowerBound (..), Version, VersionRange, VersionRangeF (..), anyVersion, asVersionIntervals, cataVersionRange, mkVersion, version0, versionNumbers)
+import Text.PrettyPrint              (Doc, comma, fsep, punctuate, text, vcat, (<+>))
 
 import qualified Data.Set                        as Set
 import qualified Distribution.Compat.CharParsing as P
@@ -71,36 +70,29 @@ class    Sep sep  where
 
     parseSep :: CabalParsing m => Proxy sep -> m a -> m [a]
 
-    describeSep :: Proxy sep -> GrammarRegex a -> GrammarRegex a
-
 instance Sep CommaVCat where
     prettySep  _ = vcat . punctuate comma
     parseSep   _ p = do
         v <- askCabalSpecVersion
         if v >= CabalSpecV2_2 then parsecLeadingCommaList p else parsecCommaList p
-    describeSep _ = reCommaList
 instance Sep CommaFSep where
     prettySep _ = fsep . punctuate comma
     parseSep   _ p = do
         v <- askCabalSpecVersion
         if v >= CabalSpecV2_2 then parsecLeadingCommaList p else parsecCommaList p
-    describeSep _ = reCommaList
 instance Sep VCat where
     prettySep _  = vcat
     parseSep   _ p = do
         v <- askCabalSpecVersion
         if v >= CabalSpecV3_0 then parsecLeadingOptCommaList p else parsecOptCommaList p
-    describeSep _ = reCommaList
 instance Sep FSep where
     prettySep _  = fsep
     parseSep   _ p = do
         v <- askCabalSpecVersion
         if v >= CabalSpecV3_0 then parsecLeadingOptCommaList p else parsecOptCommaList p
-    describeSep _ = reOptCommaList
 instance Sep NoCommaFSep where
     prettySep _   = fsep
     parseSep  _ p = many (p <* P.spaces)
-    describeSep _ = reSpacedList
 
 -- | List separated with optional commas. Displayed with @sep@, arguments of
 -- type @a@ are parsed and pretty-printed as @b@.
@@ -129,9 +121,6 @@ instance (Newtype a b, Sep sep, Parsec b) => Parsec (List sep b a) where
 
 instance (Newtype a b, Sep sep, Pretty b) => Pretty (List sep b a) where
     pretty = prettySep (Proxy :: Proxy sep) . map (pretty . (pack :: a -> b)) . unpack
-
-instance (Newtype a b, Sep sep, Described b) => Described (List sep b a) where
-    describe _ = describeSep (Proxy :: Proxy sep) (describe (Proxy :: Proxy b))
 
 --
 -- | Like 'List', but for 'Set'.
@@ -169,9 +158,6 @@ instance (Newtype a b, Ord a, Sep sep, Parsec b) => Parsec (Set' sep b a) where
 instance (Newtype a b, Sep sep, Pretty b) => Pretty (Set' sep b a) where
     pretty = prettySep (Proxy :: Proxy sep) . map (pretty . (pack :: a -> b)) . Set.toList . unpack
 
-instance (Newtype a b, Ord a, Sep sep, Described b) => Described (Set' sep b a) where
-    describe _ = describeSep (Proxy :: Proxy sep) (describe (Proxy :: Proxy b))
-
 -- | Haskell string or @[^ ,]+@
 newtype Token = Token { getToken :: String }
 
@@ -182,9 +168,6 @@ instance Parsec Token where
 
 instance Pretty Token where
     pretty = showToken . unpack
-
-instance Described Token where
-    describe _ = REUnion [reHsString, reMunch1CS csNotSpaceOrComma]
 
 -- | Haskell string or @[^ ]+@
 newtype Token' = Token' { getToken' :: String }
@@ -197,9 +180,6 @@ instance Parsec Token' where
 instance Pretty Token' where
     pretty = showToken . unpack
 
-instance Described Token' where
-    describe _ = REUnion [reHsString, reMunch1CS csNotSpace]
-
 -- | Either @"quoted"@ or @un-quoted@.
 newtype MQuoted a = MQuoted { getMQuoted :: a }
 
@@ -210,10 +190,6 @@ instance Parsec a => Parsec (MQuoted a) where
 
 instance Pretty a => Pretty (MQuoted a)  where
     pretty = pretty . unpack
-
-instance Described a => Described (MQuoted a) where
-    -- TODO: this is simplification
-    describe _ = describe ([] :: [a])
 
 -- | Version range or just version, i.e. @cabal-version@ field.
 --
@@ -302,10 +278,6 @@ instance Pretty SpecVersion where
         | csv >= CabalSpecV1_12 = text (showCabalSpecVersion csv)
         | otherwise             = text ">=" <<>> text (showCabalSpecVersion csv)
 
-instance Described SpecVersion where
-    describe _ = "3.0" -- :)
-
-
 -- | SPDX License expression or legacy license
 newtype SpecLicense = SpecLicense { getSpecLicense :: Either SPDX.License License }
 
@@ -321,9 +293,6 @@ instance Parsec SpecLicense where
 instance Pretty SpecLicense where
     pretty = either pretty pretty . unpack
 
-instance Described SpecLicense where
-    describe _ = RETodo
-
 -- | Version range or just version
 newtype TestedWith = TestedWith { getTestedWith :: (CompilerFlavor, VersionRange) }
 
@@ -336,9 +305,6 @@ instance Pretty TestedWith where
     pretty x = case unpack x of
         (compiler, vr) -> pretty compiler <+> pretty vr
 
-instance Described TestedWith where
-    describe _ = RETodo
-
 -- | Filepath are parsed as 'Token'.
 newtype FilePathNT = FilePathNT { getFilePathNT :: String }
 
@@ -349,9 +315,6 @@ instance Parsec FilePathNT where
 
 instance Pretty FilePathNT where
     pretty = showFilePath . unpack
-
-instance Described FilePathNT where
-    describe _ = describe ([] :: [Token])
 
 -------------------------------------------------------------------------------
 -- Internal
