+++ Cabal/Distribution/System.hs
@@ -50,7 +50,6 @@ import Distribution.Utils.Generic (lowercase)
 
 import Distribution.Parsec
 import Distribution.Pretty
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint as Disp
@@ -133,12 +132,7 @@ instance Pretty OS where
 instance Parsec OS where
   parsec = classifyOS Compat <$> parsecIdent
 
-instance Described OS where
-  describe _ = REUnion
-    [ fromString al
-    | os <- knownOSs
-    , al <- prettyShow os : osAliases Compat os
-    ]
+
 
 classifyOS :: ClassificationStrictness -> String -> OS
 classifyOS strictness s =
@@ -205,12 +199,6 @@ instance Pretty Arch where
 
 instance Parsec Arch where
   parsec = classifyArch Strict <$> parsecIdent
-
-instance Described Arch where
-  describe _ = REUnion
-    [ fromString (prettyShow arch)
-    | arch <- knownArches
-    ]
 
 classifyArch :: ClassificationStrictness -> String -> Arch
 classifyArch strictness s =
