+++ Cabal/Distribution/Types/BuildType.hs
@@ -13,7 +13,6 @@ import Distribution.Compat.Prelude
 import Distribution.CabalSpecVersion (CabalSpecVersion (..))
 import Distribution.Pretty
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint as Disp
@@ -54,6 +53,3 @@ instance Parsec BuildType where
               return Custom
           else fail ("unknown build-type: '" ++ name ++ "'")
       _           -> fail ("unknown build-type: '" ++ name ++ "'")
-      
-instance Described BuildType where
-    describe _ = REUnion ["Simple","Configure","Custom","Make","Default"]
