+++ Cabal/Distribution/Types/ExecutableScope.hs
@@ -11,7 +11,6 @@ import Distribution.Compat.Prelude
 
 import Distribution.Pretty
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint as Disp
@@ -28,9 +27,6 @@ instance Parsec ExecutableScope where
     parsec = P.try pub <|> pri where
         pub = ExecutablePublic  <$ P.string "public"
         pri = ExecutablePrivate <$ P.string "private"
-
-instance Described ExecutableScope where
-    describe _ = REUnion ["public","private"]
 
 instance Binary ExecutableScope
 instance Structured ExecutableScope
