+++ Cabal/Distribution/Types/ForeignLibType.hs
@@ -14,7 +14,6 @@ import Distribution.PackageDescription.Utils
 
 import Distribution.Pretty
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint as Disp
@@ -42,9 +41,6 @@ instance Parsec ForeignLibType where
       "native-shared" -> ForeignLibNativeShared
       "native-static" -> ForeignLibNativeStatic
       _               -> ForeignLibTypeUnknown
-
-instance Described ForeignLibType where
-  describe _ = REUnion ["native-shared","native-static"]
 
 instance Binary ForeignLibType
 instance Structured ForeignLibType
