+++ Cabal/Distribution/Types/LibraryVisibility.hs
@@ -11,7 +11,6 @@ import Prelude ()
 
 import Distribution.Parsec
 import Distribution.Pretty
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint                as Disp
@@ -39,9 +38,6 @@ instance Parsec LibraryVisibility where
       "public"  -> return LibraryVisibilityPublic
       "private" -> return LibraryVisibilityPrivate
       _         -> fail $ "Unknown visibility: " ++ name
-
-instance Described LibraryVisibility where
-  describe _ = REUnion ["public","private"]
 
 instance Binary LibraryVisibility
 instance Structured LibraryVisibility
