+++ Cabal/Distribution/Types/Mixin.hs
@@ -10,7 +10,6 @@ import Prelude ()
 
 import Text.PrettyPrint ((<+>))
 
-import Distribution.FieldGrammar.Described
 import Distribution.Parsec
 import Distribution.Pretty
 import Distribution.Types.IncludeRenaming
@@ -36,6 +35,3 @@ instance Parsec Mixin where
         P.spaces
         incl <- parsec
         return (Mixin mod_name incl)
-
-instance Described Mixin where
-    describe _ = RETodo
