+++ Cabal/Distribution/Types/PackageName.hs
@@ -14,7 +14,6 @@ import Distribution.Utils.ShortText
 import qualified Text.PrettyPrint as Disp
 import Distribution.Pretty
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 
 -- | A package name.
 --
@@ -72,7 +71,3 @@ instance Parsec PackageName where
 
 instance NFData PackageName where
     rnf (PackageName pkg) = rnf pkg
-
-instance Described PackageName where
-  describe _ = reUnqualComponent
-
