+++ Cabal/Distribution/Types/SourceRepo.hs
@@ -19,7 +19,6 @@ import Distribution.Utils.Generic (lowercase)
 
 import Distribution.Pretty
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint as Disp
@@ -177,9 +176,6 @@ classifyRepoKind name = case lowercase name of
 
 instance Parsec RepoType where
   parsec = classifyRepoType <$> P.munch1 isIdent
-
-instance Described RepoType where
-  describe _ = reMunch1CS $ csAlphaNum <> csChar '_' <> csChar '-'
 
 instance Pretty RepoType where
   pretty (OtherRepoType other) = Disp.text other
