+++ Cabal/Distribution/Types/UnqualComponentName.hs
@@ -11,7 +11,6 @@ import Distribution.Utils.ShortText
 import Prelude ()
 
 import Distribution.Parsec
-import Distribution.FieldGrammar.Described
 import Distribution.Pretty
 import Distribution.Types.PackageName
 
@@ -61,9 +60,6 @@ instance Pretty UnqualComponentName where
 
 instance Parsec UnqualComponentName where
   parsec = mkUnqualComponentName <$> parsecUnqualComponentName
-
-instance Described UnqualComponentName where
-  describe _ = RETodo
 
 instance NFData UnqualComponentName where
   rnf (UnqualComponentName pkg) = rnf pkg
