+++ Cabal/Distribution/Types/VersionRange/Internal.hs
@@ -39,11 +39,10 @@ import Distribution.Types.Version
 import Prelude ()
 
 import Distribution.CabalSpecVersion
-import Distribution.FieldGrammar.Described
 import Distribution.Parsec
 import Distribution.Pretty
-import Distribution.Utils.Generic          (unsnoc)
-import Text.PrettyPrint                    ((<+>))
+import Distribution.Utils.Generic    (unsnoc)
+import Text.PrettyPrint              ((<+>))
 
 import qualified Distribution.Compat.CharParsing as P
 import qualified Distribution.Compat.DList       as DList
@@ -319,33 +318,6 @@ prettyVersionRange16 vr = prettyVersionRange vr
 --
 instance Parsec VersionRange where
     parsec = askCabalSpecVersion >>= versionRangeParser versionDigitParser
-
-instance Described VersionRange where
-    describe _ = RERec "version-range" $ REUnion
-        [ "=="  <> RESpaces <> ver
-        , ">"   <> RESpaces <> ver
-        , "<"   <> RESpaces <> ver
-        , "<="  <> RESpaces <> ver
-        , ">="  <> RESpaces <> ver
-        , "^>=" <> RESpaces <> ver
-
-        -- ==0.1.*
-        , "==" <> RESpaces <> wildVer
-
-        , reVar0 <> RESpaces  <> "||" <> RESpaces <> reVar0
-        , reVar0 <> RESpaces  <> "&&" <> RESpaces <> reVar0
-        , "(" <> RESpaces <> reVar0  <> RESpaces <> ")"
-
-        -- == { 0.1.2 }
-        -- silly haddock: ^>= { 0.1.2, 3.4.5 }
-        , "=="  <> RESpaces <> verSet
-        , "^>=" <> RESpaces <> verSet
-        ]
-      where
-        ver'    = describe (Proxy :: Proxy Version)
-        ver     = RENamed "version" ver'
-        wildVer = ver' <> ".*"
-        verSet  = "{" <> RESpaces <> REMunch1 reSpacedComma ver <> RESpaces <> "}"
 
 -- | 'VersionRange' parser parametrised by version digit parser
 --
