+++ cabal-install/Distribution/Client/IndexUtils/Timestamp.hs
@@ -30,8 +30,6 @@ import Data.Time.Clock.POSIX (posixSecondsToUTCTime, utcTimeToPOSIXSeconds)
 import Distribution.Parsec   (Parsec (..))
 import Distribution.Pretty   (Pretty (..))
 
-import Distribution.FieldGrammar.Described
-
 import qualified Codec.Archive.Tar.Entry         as Tar
 import qualified Distribution.Compat.CharParsing as P
 import qualified Text.PrettyPrint                as Disp
@@ -102,119 +100,6 @@ instance Structured Timestamp
 
 instance Pretty Timestamp where
     pretty = Disp.text . showTimestamp
-
-instance Described Timestamp where
-    describe _ =  REUnion
-        [ posix
-        , utc
-        ]
-      where
-        posix = reChar '@' <> reMunch1CS "0123456789"
-        utc   = RENamed "date" date <> reChar 'T' <> RENamed "time" time <> reChar 'Z'
-
-        date = REOpt digit <> REUnion
-            [ leapYear   <> reChar '-' <> leapMD
-            , commonYear <> reChar '-' <> commonMD
-            ]
-
-        -- leap year: either
-        -- * divisible by 400
-        -- * not divisible by 100 and divisible by 4
-        leapYear = REUnion
-            [ div4           <> "00"
-            , digit <> digit <> div4not0
-            ]
-
-        -- common year: either
-        -- * not divisible by 400 but divisible by 100
-        -- * not divisible by 4
-        commonYear = REUnion
-            [ notDiv4        <> "00"
-            , digit <> digit <> notDiv4
-            ]
-
-        div4 = REUnion
-            [ "0" <> reChars "048"
-            , "1" <> reChars "26"
-            , "2" <> reChars "048"
-            , "3" <> reChars "26"
-            , "4" <> reChars "048"
-            , "5" <> reChars "26"
-            , "6" <> reChars "048"
-            , "7" <> reChars "26"
-            , "8" <> reChars "048"
-            , "9" <> reChars "26"
-            ]
-
-        div4not0 = REUnion
-            [ "0" <> reChars "48" -- no zero
-            , "1" <> reChars "26"
-            , "2" <> reChars "048"
-            , "3" <> reChars "26"
-            , "4" <> reChars "048"
-            , "5" <> reChars "26"
-            , "6" <> reChars "048"
-            , "7" <> reChars "26"
-            , "8" <> reChars "048"
-            , "9" <> reChars "26"
-            ]
-
-        notDiv4 = REUnion
-            [ "0" <> reChars "1235679"
-            , "1" <> reChars "01345789"
-            , "2" <> reChars "1235679"
-            , "3" <> reChars "01345789"
-            , "4" <> reChars "1235679"
-            , "5" <> reChars "01345789"
-            , "6" <> reChars "1235679"
-            , "7" <> reChars "01345789"
-            , "8" <> reChars "1235679"
-            , "9" <> reChars "01345789"
-            ]
-
-        leapMD = REUnion
-            [ jan, fe', mar, apr, may, jun, jul, aug, sep, oct, nov, dec ]
-
-        commonMD = REUnion
-            [ jan, feb, mar, apr, may, jun, jul, aug, sep, oct, nov, dec ]
-
-        jan = "01-" <> d31
-        feb = "02-" <> d28
-        fe' = "02-" <> d29
-        mar = "03-" <> d31
-        apr = "04-" <> d30
-        may = "05-" <> d31
-        jun = "06-" <> d30
-        jul = "07-" <> d31
-        aug = "08-" <> d31
-        sep = "09-" <> d30
-        oct = "10-" <> d31
-        nov = "11-" <> d30
-        dec = "12-" <> d31
-
-        d28 = REUnion
-            [ "0" <> digit1, "1" <> digit, "2" <> reChars "012345678" ]
-        d29 = REUnion
-            [ "0" <> digit1, "1" <> digit, "2" <> digit ]
-        d30 = REUnion
-            [ "0" <> digit1, "1" <> digit, "2" <> digit, "30" ]
-        d31 = REUnion
-            [ "0" <> digit1, "1" <> digit, "2" <> digit, "30", "31" ]
-
-        time = ho <> reChar ':' <> minSec <> reChar ':' <> minSec
-
-        -- 0..23
-        ho = REUnion
-            [ "0" <> digit
-            , "1" <> digit
-            , "2" <> reChars "0123"
-            ]
-
-        -- 0..59
-        minSec = reChars "012345" <> digit
-
-        digit  = reChars "0123456789"
-        digit1 = reChars  "123456789"
 
 instance Parsec Timestamp where
     parsec = parsePosix <|> parseUTC
