+++ cabal-install/cabal-install.cabal.pp
@@ -471,6 +471,7 @@ Test-Suite unit-tests
     UnitTests.Distribution.Client.ArbitraryInstances
     UnitTests.Distribution.Client.BuildReport
     UnitTests.Distribution.Client.Described
+    UnitTests.Distribution.Client.DescribedInstances
     UnitTests.Distribution.Client.Targets
     UnitTests.Distribution.Client.FileMonitor
     UnitTests.Distribution.Client.Get
@@ -502,6 +503,7 @@ Test-Suite unit-tests
         cabal-lib-client,
         cabal-install-solver-dsl,
         Cabal,
+        Cabal-described,
         Cabal-tree-diff,
         Cabal-QuickCheck,
         containers,
@@ -513,7 +515,6 @@ Test-Suite unit-tests
         tar,
         time,
         zlib,
-        rere >=0.1 && <0.2,
         network-uri < 2.6.2.0,
         network,
         tasty >= 1.2.3 && <1.3,
