+++ cabal.project.buildinfo
@@ -1,4 +1,5 @@
 packages: Cabal/
+packages: Cabal/Cabal-described
 packages: buildinfo-reference-generator/
 tests: False
 optimization: False
