+++ cabal.project.validate
@@ -3,6 +3,7 @@ tests: True
 
 packages: Cabal/Cabal-QuickCheck/
 packages: Cabal/Cabal-tree-diff
+packages: Cabal/Cabal-described
 
 write-ghc-environment-files: never
